requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
psutil>=5.9.5
plexapi>=4.10.0
aiohttp>=3.9.0

